Removed Gulp and added Rollup.
- Tweak: Public assets (CSS & JS) path.
- Tweak: Public assets (CSS & JS) file names.
- Dev: Removed Gulp and added [Rollup](https://rollupjs.org/) to compile the public assets.
