+++ .editorconfig
@@ -3,7 +3,7 @@ root = true
 
 [*]
 indent_style = tab
-indent_size = 4
+indent_size = 2
 end_of_line = lf
 charset = utf-8
 trim_trailing_whitespace = true
