+++ package.json
@@ -1,46 +1,45 @@
 {
-  "name": "addonify-wishlist",
-  "version": "0.1.0",
-  "description": "Addonify WooCommerce Wishlist is a powerful toolkit which adds a wishlist functionality to your e-commerce shop.",
-  "license": "GPLv3",
-  "private": true,
-  "author": "Addonify",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "public": "gulp assets",
-    "zip": "vite build && gulp zip"
-  },
-  "dependencies": {
-    "@fontsource/inter": "^5.0.5",
-    "@fontsource/manrope": "^5.0.5",
-    "element-plus": "^2.3.8",
-    "pinia": "^2.0.33",
-    "vue": "^3.2.47",
-    "vue-router": "^4.1.6"
-  },
-  "devDependencies": {
-    "@kucrut/vite-for-wp": "^0.5.2",
-    "@vitejs/plugin-vue": "^4.0.0",
-    "autoprefixer": "^10.4.14",
-    "cssnano": "^5.1.15",
-    "gulp": "^4.0.2",
-    "gulp-concat": "^2.6.1",
-    "gulp-notify": "^4.0.0",
-    "gulp-plumber": "^1.2.1",
-    "gulp-postcss": "^9.0.1",
-    "gulp-rename": "^2.0.0",
-    "gulp-rtlcss": "^2.0.0",
-    "gulp-sass": "^5.1.0",
-    "gulp-shell": "^0.8.0",
-    "gulp-sourcemaps": "^3.0.0",
-    "gulp-uglify": "^3.0.2",
-    "gulp-zip": "^5.1.0",
-    "sass": "^1.59.2",
-    "typescript": "^4.9.5",
-    "unplugin-auto-import": "^0.15.1",
-    "unplugin-vue-components": "^0.24.1",
-    "vite": "^4.1.4",
-    "vue-tsc": "^1.2.0"
-  }
+	"name": "addonify-wishlist",
+	"version": "0.1.0",
+	"description": "Addonify WooCommerce Wishlist is a powerful toolkit which adds a wishlist functionality to your e-commerce shop.",
+	"license": "GPLv3",
+	"private": true,
+	"author": "Addonify",
+	"type": "module",
+	"scripts": {
+		"public": "rollup -c -w",
+		"dev:admin": "vite",
+		"build:admin": "vite build"
+	},
+	"dependencies": {
+		"@fontsource/inter": "^5.0.5",
+		"@fontsource/manrope": "^5.0.5",
+		"element-plus": "^2.3.8",
+		"pinia": "^2.0.33",
+		"vue": "^3.2.47",
+		"vue-router": "^4.1.6"
+	},
+	"devDependencies": {
+		"@babel/preset-env": "^7.26.0",
+		"@kucrut/vite-for-wp": "^0.5.2",
+		"@rollup/plugin-alias": "^5.1.1",
+		"@rollup/plugin-babel": "^6.0.4",
+		"@rollup/plugin-commonjs": "^28.0.2",
+		"@rollup/plugin-node-resolve": "^16.0.0",
+		"@rollup/plugin-terser": "^0.4.4",
+		"@vitejs/plugin-vue": "^4.0.0",
+		"autoprefixer": "^10.4.14",
+		"babel": "^6.23.0",
+		"cssnano": "^5.1.15",
+		"gulp-uglify": "^3.0.2",
+		"postcss-rtlcss": "^5.6.0",
+		"rollup": "^4.30.1",
+		"rollup-plugin-scss": "^4.0.1",
+		"sass": "^1.75.0",
+		"typescript": "^4.9.5",
+		"unplugin-auto-import": "^0.15.1",
+		"unplugin-vue-components": "^0.24.1",
+		"vite": "^4.1.4",
+		"vue-tsc": "^1.2.0"
+	}
 }
