+++ public/assets/src/import.common.js
@@ -1,0 +1,4 @@
+/**
+* Import common.
+*/
+import "./js/common";
