+++ public/assets/src/import.global.js
@@ -1,0 +1,9 @@
+/**
+* Import scss.
+*/
+import "./scss/index.scss";
+
+/**
+* Import other global scripts.
+*/
+import "./js/perfect-scrollbar";
