+++ public/assets/src/import.guest.js
@@ -1,0 +1,4 @@
+/**
+* Import guest.
+*/
+import "./js/guest";
