+++ public/assets/src/import.private.js
@@ -1,0 +1,4 @@
+/**
+* Import private.
+*/
+import "./js/private";
