+++ public/assets/src/scss/abstracts/_mixins.scss
@@ -1,4 +1,4 @@
-/* -------------------------------------------------------------------------- 
+/* --------------------------------------------------------------------------
 	Mixin: Better Font Rendering
   -------------------------------------------------------------------------- */
 
@@ -13,7 +13,7 @@
     @return #{stripUnit($px) / stripUnit($html-font-size)}rem;
 }
 
-/* -------------------------------------------------------------------------- 
+/* --------------------------------------------------------------------------
 	Mixin: Flex box
   -------------------------------------------------------------------------- */
 
@@ -25,7 +25,7 @@
 }
 
 
-/* -------------------------------------------------------------------------- 
+/* --------------------------------------------------------------------------
 	Mixin: Grid
   -------------------------------------------------------------------------- */
 
@@ -59,7 +59,7 @@
   }
 }
 
-/* -------------------------------------------------------------------------- 
+/* --------------------------------------------------------------------------
 	Mixin: Visibility
   -------------------------------------------------------------------------- */
 
@@ -76,33 +76,22 @@
 }
 
 
-/* -------------------------------------------------------------------------- 
+/* --------------------------------------------------------------------------
 	Mixin: Animations & transition
   -------------------------------------------------------------------------- */
-
-@mixin transition_delay () {
-
-    transition: all 0.5s ease;
-}
-
-@mixin transition_delay_none () {
-
-    transition: none;
-}
-
 @mixin scale () {
 
     transform: scale(1.1);
-    @include transition_delay();
+    transition: all 0.5s ease;
 }
 
 
-/* -------------------------------------------------------------------------- 
+/* --------------------------------------------------------------------------
 	Mixin: Position
   -------------------------------------------------------------------------- */
 
 @mixin position ( $top, $bottom, $left, $right ) {
-	
+
 	top: $top;
 	bottom: $bottom;
 	left: $left;
@@ -111,14 +100,14 @@
 }
 
 @mixin position_topnleft () {
-	
+
 	top: 50%;
 	left: 50%;
     transform: translateX(-50%) translateY(-50%);
 }
 
 
-/* -------------------------------------------------------------------------- 
+/* --------------------------------------------------------------------------
 	Mixin: Overlay
   -------------------------------------------------------------------------- */
 
@@ -133,7 +122,7 @@
 }
 
 
-/* -------------------------------------------------------------------------- 
+/* --------------------------------------------------------------------------
 	Mixin: Drop shadow
   -------------------------------------------------------------------------- */
 
@@ -152,7 +141,7 @@
     box-shadow: 0px 0px 65px rgba(0, 0, 0, 0.2);
 }
 
-/* -------------------------------------------------------------------------- 
+/* --------------------------------------------------------------------------
 	Mixin: Clearfix
   -------------------------------------------------------------------------- */
 
@@ -171,7 +160,7 @@
 	}
 }
 
-/* -------------------------------------------------------------------------- 
+/* --------------------------------------------------------------------------
 	Mixin: Breakpoints
   -------------------------------------------------------------------------- */
