+++ public/assets/src/scss/layout/_wishlist-page.scss
@@ -1,4 +1,4 @@
-/* -------------------------------------------------------------------------- 
+/* --------------------------------------------------------------------------
 	#. Wishlist page
   -------------------------------------------------------------------------- */
 
@@ -25,7 +25,7 @@
 
         td.price {
 
-            font-weight: bold; 
+            font-weight: bold;
 
             ins {
 
@@ -121,7 +121,7 @@
         bottom: 0;
         z-index: 100;
         background-color: rgba(255, 255, 255, 1);
-        @include transition_delay();
+        transition: all 0.5s ease;
 
         svg {
 
@@ -137,8 +137,8 @@
     /**
     *
     * Notice
-    * 
-    */ 
+    *
+    */
 
     #addonify-wishlist-notice  {
 
@@ -156,7 +156,7 @@
     /**
     *
     * Empty wishlist
-    * 
+    *
     */
 
     #addonify-wishlist-empty {
@@ -166,7 +166,7 @@
 }
 
 
-/* -------------------------------------------------------------------------- 
+/* --------------------------------------------------------------------------
 	#. Media Query
   -------------------------------------------------------------------------- */
 
@@ -206,7 +206,7 @@
 
                             padding-bottom: 0;
                             border-bottom: none
-                        } 
+                        }
                     }
 
                     td.remove {
