+++ public/assets/src/scss/layout/_wishlist-sidebar.scss
@@ -1,4 +1,4 @@
-/* -------------------------------------------------------------------------- 
+/* --------------------------------------------------------------------------
 	#. Wishlist sidebar panel
   -------------------------------------------------------------------------- */
 
@@ -20,8 +20,8 @@
     padding: 30px 20px 20px 20px;
     z-index: var(--sidebar-zindex);
     background-color: var(--adfy_wishlist_sidebar_modal_bg_color);
+    transition: all 0.5s ease;
     @include large_drop_shadow();
-    @include transition_delay();
     @include visibility_visible();
     @include breakpoint(mobile) {
 
@@ -87,7 +87,7 @@
                     transform: none;
                     text-decoration: none;
                 }
-            }   
+            }
         }
 
         #addonify-wishlist-sidebar-form {
@@ -315,7 +315,7 @@
             border-radius: 30px;
             box-shadow: none;
             text-decoration: none;
-            @include transition_delay();
+            transition: all 0.5s ease;
 
             .icon {
 
@@ -344,7 +344,7 @@
     /**
     *
     * Notice
-    * 
+    *
     */
 
     #addonify-wishlist-undo-deleted-product  {
@@ -425,7 +425,7 @@
     background: transparent;
     background-color: var(--adfy_wishlist_sidebar_modal_toggle_btn_bg_color);
     z-index: var(--zindex);
-    @include transition_delay();
+    transition: all 0.5s ease;
 
     .button-icon,
     .button-label {
@@ -440,7 +440,7 @@
     }
 
     .button-icon {
-        
+
         transform: rotate(-90deg);
 
         svg {
@@ -486,7 +486,7 @@
     opacity: 0;
     visibility: hidden;
     background-color: var(--adfy_wishlist_sidebar_modal_overlay_bg_color);
-    @include transition_delay();
+    transition: all 0.5s ease;
 }
 
 #addonify-wishlist-sticky-sidebar-container #addonify-wishlist_spinner {
@@ -504,7 +504,7 @@
     z-index: 100;
     backdrop-filter: blur(5px);
     background-color: var(--adfy_wishlist_sidebar_modal_spinner_bg_color);
-    @include transition_delay();
+    transition: all 0.5s ease;
 
     svg {
 
@@ -517,7 +517,7 @@
     }
 }
 
-/* -------------------------------------------------------------------------- 
+/* --------------------------------------------------------------------------
 	#. Visibility & Alignment
   -------------------------------------------------------------------------- */
 
@@ -568,7 +568,7 @@
 
             left: auto;
             right: 0;
-        }        
+        }
     }
 }
